And a very rough out of a generic client middleman
It is to live between the the socket handling code (netstrings and such)
and the code that will actually control a ship, present a list of
characters, or control a player on the ground.  Currently a bit bound to
the protbufs as those need to be encoded to binary before the socket
process and send them, but at least it doesn't have to be netstringed.