Enforce transactions on many api things.
The real reason to do this is to force transaction required functions
to all run within the same pid. Some backends (mnesia) require transations
not cross pid bounderies (it uses the process dictionary to track
transactions). This change does the same to keep track of the callback
module. This has the advantage of requiring fewer trips to the pre_data
process, so pragmatism beats the 'pdict is dirty' purity.