Adding holder items so a git checkout creates ebins.
rebar deps verification was failing for sub-apps. This is because rebar
looks for an *.app file in the ebin directory before looking in the src
directory for a *.app.src. If rebar doesn't even find the ebin directory
(possibly because this is a brand new checkout), then it skips the second
step, assuming it is not an otp application. By placing the ebin directory
there, but not putting the *.app file there, rebar will carry on checking
the src directory, and also create the appropriate app file when needed.